refactor: refactor deleteGuildRole util and tests in separate file
