+++ src/utils/guildRole.ts
@@ -23,31 +23,6 @@ import { GuildRole, Role } from "../typeDefinitions/role.types";
 import createDiscordHeaders from "./createDiscordHeaders";
 import { sleep } from "./sleep";
 
-export async function deleteGuildRole(
-  env: env,
-  roleId: string,
-  reason?: string
-) {
-  const deleteGuildRoleUrl = `${DISCORD_BASE_URL}/guilds/${env.DISCORD_GUILD_ID}/roles/${roleId}`;
-  const headers: HeadersInit = createDiscordHeaders({
-    token: env.DISCORD_TOKEN,
-    reason: reason,
-  });
-  try {
-    const response = await fetch(deleteGuildRoleUrl, {
-      method: "DELETE",
-      headers,
-    });
-    if (response.ok) {
-      return ROLE_REMOVED;
-    } else {
-      return INTERNAL_SERVER_ERROR;
-    }
-  } catch (err) {
-    return INTERNAL_SERVER_ERROR;
-  }
-}
-
 export async function createGuildRole(
   body: createNewRole,
   env: env,
